ckanext-ilrimetadata: Improve error handling
Python subprocess.call doesn't seem to return anything if the process
exits with non-zero code. For example, if mysqltofile is not present
on the server there is nothing printed in the logs or the browser. I
decided to try/except on the subprocess.call and set the result to a
non-zero value manually if needed, then abort with 503 so that there
is at least a chance of detecting this if looking at the server logs
or the web browser network log.